+++ test/deploymentUtils.js
@@ -170,6 +170,39 @@ async function deployCMTATStandaloneWithParameter (
   return cmtat
 }
 
+async function deployCMTATHolderListStandalone (forwarder, admin, deployerAddress) {
+  const cmtat = await ethers.deployContract('CMTATStandaloneHolderList', [
+    forwarder,
+    admin,
+    ['CMTA Token', 'CMTAT', DEPLOYMENT_DECIMAL],
+    ['CMTAT_ISIN', TERMS, 'CMTAT_info'],
+    [ZERO_ADDRESS]
+  ])
+  return cmtat
+}
+
+async function deployCMTATHolderListProxy (forwarder, admin, deployerAddress) {
+  const ETHERS_CMTAT_PROXY_FACTORY = await ethers.getContractFactory(
+    'CMTATUpgradeableHolderList'
+  )
+  const ETHERS_CMTAT_PROXY = await upgrades.deployProxy(
+    ETHERS_CMTAT_PROXY_FACTORY,
+    [
+      admin,
+      ['CMTA Token', 'CMTAT', DEPLOYMENT_DECIMAL],
+      ['CMTAT_ISIN', TERMS, 'CMTAT_info'],
+      [ZERO_ADDRESS]
+    ],
+    {
+      initializer: 'initialize',
+      constructorArgs: [forwarder],
+      from: deployerAddress,
+      unsafeAllow: ['missing-initializer']
+    }
+  )
+  return ETHERS_CMTAT_PROXY
+}
+
 async function deployCMTATERC1363Proxy (forwarder, admin, deployerAddress) {
   // Ref: https://forum.openzeppelin.com/t/upgrades-hardhat-truffle5/30883/3
   const ETHERS_CMTAT_PROXY_FACTORY = await ethers.getContractFactory(
@@ -439,6 +472,8 @@ module.exports = {
   deployCMTATDebtEngineProxy,
   deployCMTATERC1363Proxy,
   deployCMTATERC1363Standalone,
+  deployCMTATHolderListProxy,
+  deployCMTATHolderListStandalone,
   deployCMTATPermitStandalone,
   deployCMTATSnapshotStandalone,
   deployCMTATERC7551Proxy,